feat(mdx): Pocket実機シーケンス検証を追加する
変更:
- 自作FM-only MDXをparseからDeviceSchedulerまで通す固定ハッシュプローブを追加
- RISC-V/desktop/hostの同値検証とM3安全配置パッケージを追加
- 実機表示、成果物ハッシュ、検証手順を文書化

理由:
- M3のホスト受け入れ結果をPocket上でも同一経路で確認するため

検証:
- pwsh -NoProfile -File tools/host-verify.ps1: 30/30 PASS
- pwsh -NoProfile -File tools/pocket-toolchain-verify.ps1: PASS

+++ core/runtime/engines/mdx/mdx_ym2151_router.cpp
@@ -119,7 +119,7 @@ std::uint16_t compute_gate_ticks(const Ym2151ChannelState& state,
     return static_cast<std::uint16_t>(
         (static_cast<std::uint16_t>(state.gate_parameter) * raw) / 8U + 1U);
   }
-  const auto sum = static_cast<std::uint16_t>(state.gate_parameter) + raw;
+  const auto sum = static_cast<std::uint32_t>(state.gate_parameter) + raw;
   return sum >= 0x100U ? static_cast<std::uint16_t>((sum & 0xffU) + 1U) : 1U;
 }
